+++ Lib-UAV/src/lib/uav/struct/constants/TypeReplanner.java
@@ -7,13 +7,14 @@ package lib.uav.struct.constants;
  */
 public class TypeReplanner {
    
-    public static final String DE4S          = "DE4s";
-    public static final String GH4S          = "GH4s";
-    public static final String GA4S          = "GA4s";
-    public static final String MPGA4S        = "MPGA4s";
-    public static final String MS4S          = "MS4s";
-    public static final String GA_GA_4S      = "GA-GA-4s";
-    public static final String GA_GH_4S      = "GA-GH-4s";
-    public static final String PRE_PLANNED4s = "Pre-Planned4s";
+    public static final String DE4S               = "DE4s";
+    public static final String GH4S               = "GH4s";
+    public static final String GA4S               = "GA4s";
+    public static final String MPGA4S             = "MPGA4s";
+    public static final String MS4S               = "MS4s";
+    public static final String GA_GA_4S           = "GA-GA-4s";
+    public static final String GA_GH_4S           = "GA-GH-4s";
+    public static final String PRE_PLANNED4s      = "Pre-Planned4s";
+    public static final String G_PATH_REPLANNER4M = "G_PATH_REPLANNER4m";
     
 }