+++ Modules-IFA/G-Path-Replanner4s/exec-replanner.sh
@@ -1,0 +1,16 @@
+#!/bin/bash
+#Author: Jesimar da Silva Arantes
+#Date: 05/10/2018
+#Last Update: 05/10/2018
+#Description: Script that runs the path replanner: G-Path-Replanner4s.
+#Descrição: Script que executa o replanejador de rotas: G-Path-Replanner4s.
+
+#Abaixo deve ser definida a forma de execução desse código usando o shellscript.
+
+#Por exemplo: caso seu código esteja em Java use a linha abaixo.
+java -jar nome_path_replanner4s.jar > output-simulation.log
+
+#Por exemplo: caso seu código esteja em C use a linha abaixo.
+./nome_path_replanner4s > output-simulation.log
+
+#Caso seu código esteja em outra linguagem de programação diferente defina abaixo a forma de execução desse código através do terminal.