+++ UAV-GCS/src/uav/gcs/plotmap/PanelPlotMission.java
@@ -389,7 +389,7 @@ public class PanelPlotMission extends sPanelDraw {
             });
         }
         if (config.getSystemExecMOSA().equals(TypeSystemExecMOSA.PLANNER)
-                && config.getTypePlanner().equals(TypePlanner.PATH_PLANNER4M)) {
+                && config.getTypePlanner().equals(TypePlanner.G_PATH_PLANNER4M)) {
             Executors.newSingleThreadExecutor().execute(new Runnable() {
                 @Override
                 public void run() {